Uploading code for utilizing more naive models
Includes code for maximum likelihood classifier and an application of a QDA classifier. Both these classifiers are used on uniform distributions, where there are two classes 1 and 2. Class 2 contains the data of class 1, but shifted X and Y by an alpha value. The maximum likelihood implementation experiments with changing the alpha value while keeping a 50/50 prior probability split. The QDA implementation experiments with changing the prior probabilities of each class.