Uploading code and data for SVM and RNF code
Python file performs some predictions over the data using Support Vector Machines and Random Forest, and compares performance. There are also features to view how the performance of each model changes as the parameters are adjusted (i.e. number of support vectors for SVM or number of estimators for RNF)